requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==5.0.4
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
sqlparse==0.4.4